Refactor v2 parser error type
